+++ src/SMAPI.Mods.ErrorHandler/Patches/EventPatcher.cs
@@ -1,11 +1,7 @@
 using System;
 using System.Diagnostics.CodeAnalysis;
-#if HARMONY_2
 using HarmonyLib;
-#else
-using Harmony;
-#endif
-using StardewModdingAPI.Framework.Patching;
+using StardewModdingAPI.Internal.Patching;
 using StardewValley;
 
 namespace StardewModdingAPI.Mods.ErrorHandler.Patches
@@ -14,7 +10,7 @@ namespace StardewModdingAPI.Mods.ErrorHandler.Patches
     /// <remarks>Patch methods must be static for Harmony to work correctly. See the Harmony documentation before renaming patch arguments.</remarks>
     [SuppressMessage("ReSharper", "InconsistentNaming", Justification = "Argument names are defined by Harmony and methods are named for clarity.")]
     [SuppressMessage("ReSharper", "IdentifierTypo", Justification = "Argument names are defined by Harmony and methods are named for clarity.")]
-    internal class EventPatches : IHarmonyPatch
+    internal class EventPatcher : BasePatcher
     {
         /*********
         ** Fields
@@ -28,21 +24,17 @@ namespace StardewModdingAPI.Mods.ErrorHandler.Patches
         *********/
         /// <summary>Construct an instance.</summary>
         /// <param name="monitorForGame">Writes messages to the console and log file on behalf of the game.</param>
-        public EventPatches(IMonitor monitorForGame)
+        public EventPatcher(IMonitor monitorForGame)
         {
-            EventPatches.MonitorForGame = monitorForGame;
+            EventPatcher.MonitorForGame = monitorForGame;
         }
 
         /// <inheritdoc />
-#if HARMONY_2
-        public void Apply(Harmony harmony)
-#else
-        public void Apply(HarmonyInstance harmony)
-#endif
+        public override void Apply(Harmony harmony, IMonitor monitor)
         {
             harmony.Patch(
-                original: AccessTools.Method(typeof(Event), nameof(Event.LogErrorAndHalt)),
-                postfix: new HarmonyMethod(this.GetType(), nameof(EventPatches.After_Event_LogErrorAndHalt))
+                original: this.RequireMethod<Event>(nameof(Event.LogErrorAndHalt)),
+                postfix: this.GetHarmonyMethod(nameof(EventPatcher.After_LogErrorAndHalt))
             );
         }
 
@@ -52,9 +44,9 @@ namespace StardewModdingAPI.Mods.ErrorHandler.Patches
         *********/
         /// <summary>The method to call after <see cref="Event.LogErrorAndHalt"/>.</summary>
         /// <param name="e">The exception being logged.</param>
-        private static void After_Event_LogErrorAndHalt(Exception e)
+        private static void After_LogErrorAndHalt(Exception e)
         {
-            EventPatches.MonitorForGame.Log(e.ToString(), LogLevel.Error);
+            EventPatcher.MonitorForGame.Log(e.ToString(), LogLevel.Error);
         }
     }
 }