+++ src/SMAPI/Framework/ContentManagers/ModContentManager.cs
@@ -7,6 +7,7 @@ using Microsoft.Xna.Framework.Content;
 using Microsoft.Xna.Framework.Graphics;
 using StardewModdingAPI.Framework.Exceptions;
 using StardewModdingAPI.Framework.Reflection;
+using StardewModdingAPI.Internal;
 using StardewModdingAPI.Toolkit.Serialization;
 using StardewModdingAPI.Toolkit.Utilities;
 using StardewValley;
@@ -77,6 +78,8 @@ namespace StardewModdingAPI.Framework.ContentManagers
         /// <inheritdoc />
         public override T Load<T>(string assetName, LanguageCode language, bool useCache)
         {
+            // normalize key
+            bool isXnbFile = Path.GetExtension(assetName).ToLower() == ".xnb";
             assetName = this.AssertAndNormalizeAssetName(assetName);
 
             // disable caching
@@ -108,7 +111,7 @@ namespace StardewModdingAPI.Framework.ContentManagers
             try
             {
                 // get file
-                FileInfo file = this.GetModFile(assetName);
+                FileInfo file = this.GetModFile(isXnbFile ? $"{assetName}.xnb" : assetName); // .xnb extension is stripped from asset names passed to the content manager
                 if (!file.Exists)
                     throw GetContentError("the specified path doesn't exist.");