+++ src/SMAPI/Framework/ModLoading/Rewriters/Harmony1AssemblyRewriter.cs
@@ -1,4 +1,3 @@
-#if HARMONY_2
 using System;
 using HarmonyLib;
 using Mono.Cecil;
@@ -29,11 +28,11 @@ namespace StardewModdingAPI.Framework.ModLoading.Rewriters
         public override bool Handle(ModuleDefinition module, TypeReference type, Action<TypeReference> replaceWith)
         {
             // rewrite Harmony 1.x type to Harmony 2.0 type
-            if (type.Scope is AssemblyNameReference scope && scope.Name == "0Harmony" && scope.Version.Major == 1)
+            if (type.Scope is AssemblyNameReference { Name: "0Harmony" } scope && scope.Version.Major == 1)
             {
                 Type targetType = this.GetMappedType(type);
                 replaceWith(module.ImportReference(targetType));
-                this.MarkRewritten();
+                this.OnChanged();
                 this.ReplacedTypes = true;
                 return true;
             }
@@ -42,19 +41,25 @@ namespace StardewModdingAPI.Framework.ModLoading.Rewriters
         }
 
         /// <inheritdoc />
-        public override bool Handle(ModuleDefinition module, ILProcessor cil, Instruction instruction, Action<Instruction> replaceWith)
+        public override bool Handle(ModuleDefinition module, ILProcessor cil, Instruction instruction)
         {
             // rewrite Harmony 1.x methods to Harmony 2.0
             MethodReference methodRef = RewriteHelper.AsMethodReference(instruction);
             if (this.TryRewriteMethodsToFacade(module, methodRef))
+            {
+                this.OnChanged();
                 return true;
+            }
 
             // rewrite renamed fields
             FieldReference fieldRef = RewriteHelper.AsFieldReference(instruction);
             if (fieldRef != null)
             {
                 if (fieldRef.DeclaringType.FullName == "HarmonyLib.HarmonyMethod" && fieldRef.Name == "prioritiy")
+                {
                     fieldRef.Name = nameof(HarmonyMethod.priority);
+                    this.OnChanged();
+                }
             }
 
             return false;
@@ -64,6 +69,13 @@ namespace StardewModdingAPI.Framework.ModLoading.Rewriters
         /*********
         ** Private methods
         *********/
+        /// <summary>Update the mod metadata when any Harmony 1.x code is migrated.</summary>
+        private void OnChanged()
+        {
+            this.MarkRewritten();
+            this.MarkFlag(InstructionHandleResult.DetectedGamePatch);
+        }
+
         /// <summary>Rewrite methods to use Harmony facades if needed.</summary>
         /// <param name="module">The assembly module containing the method reference.</param>
         /// <param name="methodRef">The method reference to map.</param>
@@ -73,24 +85,15 @@ namespace StardewModdingAPI.Framework.ModLoading.Rewriters
                 return false; // not Harmony (or already using Harmony 2.0)
 
             // get facade type
-            Type toType;
-            switch (methodRef?.DeclaringType.FullName)
+            Type toType = methodRef?.DeclaringType.FullName switch
             {
-                case "HarmonyLib.Harmony":
-                    toType = typeof(HarmonyInstanceFacade);
-                    break;
-
-                case "HarmonyLib.AccessTools":
-                    toType = typeof(AccessToolsFacade);
-                    break;
-
-                case "HarmonyLib.HarmonyMethod":
-                    toType = typeof(HarmonyMethodFacade);
-                    break;
-
-                default:
-                    return false;
-            }
+                "HarmonyLib.Harmony" => typeof(HarmonyInstanceFacade),
+                "HarmonyLib.AccessTools" => typeof(AccessToolsFacade),
+                "HarmonyLib.HarmonyMethod" => typeof(HarmonyMethodFacade),
+                _ => null
+            };
+            if (toType == null)
+                return false;
 
             // map if there's a matching method
             if (RewriteHelper.HasMatchingSignature(toType, methodRef))
@@ -103,18 +106,24 @@ namespace StardewModdingAPI.Framework.ModLoading.Rewriters
         }
 
         /// <summary>Get an equivalent Harmony 2.x type.</summary>
-        /// <param name="type">The Harmony 1.x method.</param>
+        /// <param name="type">The Harmony 1.x type.</param>
         private Type GetMappedType(TypeReference type)
         {
-            // main Harmony object
-            if (type.FullName == "Harmony.HarmonyInstance")
-                return typeof(Harmony);
+            return type.FullName switch
+            {
+                "Harmony.HarmonyInstance" => typeof(Harmony),
+                "Harmony.ILCopying.ExceptionBlock" => typeof(ExceptionBlock),
+                _ => this.GetMappedTypeByConvention(type)
+            };
+        }
 
-            // other objects
+        /// <summary>Get an equivalent Harmony 2.x type using the convention expected for most types.</summary>
+        /// <param name="type">The Harmony 1.x type.</param>
+        private Type GetMappedTypeByConvention(TypeReference type)
+        {
             string fullName = type.FullName.Replace("Harmony.", "HarmonyLib.");
-            string targetName = typeof(Harmony).AssemblyQualifiedName.Replace(typeof(Harmony).FullName, fullName);
+            string targetName = typeof(Harmony).AssemblyQualifiedName!.Replace(typeof(Harmony).FullName!, fullName);
             return Type.GetType(targetName, throwOnError: true);
         }
     }
 }
-#endif