+++ src/SMAPI/Properties/AssemblyInfo.cs
@@ -1,5 +1,4 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("SMAPI.Tests")]
-[assembly: InternalsVisibleTo("ErrorHandler")]
 [assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")] // Moq for unit testing